pyproject.toml
[tool.poetry]
name = "linkedin-agent"
version = "1.0.0"
description = "Production-ready AI system for intelligent LinkedIn opportunity analysis using DSPy and LLM technology"
authors = ["Sebastian Davila <your.email@example.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/yourusername/linkedin-agent"
repository = "https://github.com/yourusername/linkedin-agent"
keywords = ["ai", "llm", "dspy", "linkedin", "recruiting", "fastapi"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[tool.poetry.dependencies]
python = "^3.11"

# Web Framework
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-multipart = "^0.0.6"

# Database
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
alembic = "^1.13.1"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"

# Redis & Caching
redis = {extras = ["hiredis"], version = "^5.0.1"}
aioredis = "^2.0.1"

# AI/ML
dspy-ai = "^2.4.0"
ollama = "^0.1.6"

# Background Jobs
celery = {extras = ["redis"], version = "^5.3.4"}
flower = "^2.0.1"

# Scraping
playwright = "^1.41.0"
beautifulsoup4 = "^4.12.3"
lxml = "^5.1.0"

# HTTP Client
httpx = "^0.27.0"
aiohttp = "^3.9.1"

# Monitoring & Observability
prometheus-client = "^0.19.0"
opentelemetry-api = "^1.22.0"
opentelemetry-sdk = "^1.22.0"
opentelemetry-instrumentation-fastapi = "^0.43b0"
opentelemetry-instrumentation-sqlalchemy = "^0.43b0"
opentelemetry-instrumentation-redis = "^0.43b0"
opentelemetry-exporter-otlp = "^1.22.0"
opentelemetry-exporter-prometheus = "^0.43b0"

# Logging
structlog = "^24.1.0"
python-json-logger = "^2.0.7"

# Configuration
python-dotenv = "^1.0.0"
pyyaml = "^6.0.1"

# Utilities
python-dateutil = "^2.8.2"
pytz = "^2024.1"
tenacity = "^8.2.3"

[tool.poetry.group.dev.dependencies]
# Testing
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
pytest-env = "^1.1.3"
pytest-xdist = "^3.5.0"
pytest-timeout = "^2.2.0"
coverage = {extras = ["toml"], version = "^7.4.0"}
factory-boy = "^3.3.0"
faker = "^22.2.0"

# Load Testing
locust = "^2.20.0"

# Code Quality
black = "^24.1.0"
ruff = "^0.1.14"
mypy = "^1.8.0"
isort = "^5.13.2"

# Type Stubs
types-python-dateutil = "^2.8.19"
types-pytz = "^2024.1.0"
types-pyyaml = "^6.0.12"
types-redis = "^4.6.0"

# Pre-commit
pre-commit = "^3.6.0"

# Security
bandit = "^1.7.6"
safety = "^3.0.1"

# Documentation
mkdocs = "^1.5.3"
mkdocs-material = "^9.5.4"

# Debugging
ipython = "^8.20.0"
ipdb = "^0.13.13"

# Development Tools
watchdog = "^4.0.0"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# ==========================================
# BLACK CONFIGURATION
# ==========================================
[tool.black]
line-length = 100
target-version = ['py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | migrations
)/
'''

# ==========================================
# RUFF CONFIGURATION
# ==========================================
[tool.ruff]
line-length = 100
target-version = "py311"

select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]

ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex (we'll handle this manually)
]

exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "__pycache__",
    "migrations",
    "build",
    "dist",
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__ are ok

[tool.ruff.isort]
known-first-party = ["app"]

# ==========================================
# MYPY CONFIGURATION
# ==========================================
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true

plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "dspy.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "ollama.*"
ignore_missing_imports = true

# ==========================================
# PYTEST CONFIGURATION
# ==========================================
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
    "--strict-config",
    "--cov=app",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=html",
    "--cov-report=xml",
]

markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "slow: Slow tests",
    "skip_ci: Skip in CI",
]

filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

env = [
    "ENV=test",
    "DATABASE_URL=postgresql://test:test@localhost:5432/test_db",
    "REDIS_URL=redis://localhost:6379/1",
]

# ==========================================
# COVERAGE CONFIGURATION
# ==========================================
[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/__pycache__/*",
    "*/site-packages/*",
    "*/.venv/*",
]
branch = true

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "if typing.TYPE_CHECKING:",
    "@abstractmethod",
    "@abc.abstractmethod",
]

[tool.coverage.html]
directory = "htmlcov"

# ==========================================
# BANDIT CONFIGURATION
# ==========================================
[tool.bandit]
exclude_dirs = ["tests", "migrations"]
skips = ["B101", "B601"]

# ==========================================
# ISORT CONFIGURATION
# ==========================================
[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["app"]
skip = [".venv", "migrations"]